+++ src/core/mahjong/compact_helpers.hpp
@@ -12,11 +12,11 @@ constexpr bool _is_honor(int x) MJ_EXCEPT_WARN
     return x >= k_FirstHonorIdx; }
 constexpr bool _is_19(int x) MJ_EXCEPT_WARN
 {   MJ_ASSERT(x < k_UniqueTiles, "_is_19() called on invalid tile");
-    return x % 9 == c_num(1) || x % 9 == c_num(9) || _is_honor(x); }
+    return x % 9 == 0 || x % 9 == 8 || _is_honor(x); }
 constexpr bool _is_17(int x) MJ_EXCEPT_WARN
 {   MJ_ASSERT(x < k_UniqueTiles, "_is_17() called on invalid tile");
     MJ_ASSERT(!_is_honor(x), "_is_17() called on honor tile");
-    return x % 9 == c_num(1) || x % 9 == c_num(7); }
+    return x % 9 == 0 || x % 9 == 6; }
 
 constexpr bool _is_set(int x) noexcept { return x & 0x100; }
 constexpr bool _is_run(int x) noexcept { return !(x & 0x100); }
