+++ src/core/mahjong/scoring_base.hpp
@@ -1,6 +1,6 @@
 
 #pragma once
-#include "mahjong.hpp"
+#include "mahjong_hand.hpp"
 #include "compact_helpers.hpp"
 #include "core/utils/logging.hpp"
 #include <bitset>
@@ -18,78 +18,79 @@ constexpr U16f k_Baiman = 4000;
 constexpr U16f k_Sanbaiman = 6000;
 constexpr U16f k_Yakuman = 8000;
 
+enum YakuFlags : U64 
+{
+    f_Riichi1 = 1ull << 8,
+    f_Riichi2 = 1ull << 9,
+    f_Ippatsu = 1ull << 10,
+    f_MenTsumo = 1ull << 11,
+    f_Pinfu = 1ull << 12,
+    f_Ipeikou = 1ull << 13,
+    f_Haitei = 1ull << 14,
+    f_Houtei = 1ull << 15,
+    f_Rinshan = 1ull << 16,
+    f_Chankan = 1ull << 17,
+    f_Tanyao = 1ull << 18,
+    f_Hatsu = 1ull << 19,
+    f_Chun = 1ull << 20,
+    f_Haku = 1ull << 21,
+    f_EastP = 1ull << 22,
+    f_EastS = 1ull << 23,
+    f_SouthP = 1ull << 24,
+    f_SouthS = 1ull << 25,
+    f_WestS = 1ull << 26,
+    f_NorthS = 1ull << 27,
+    f_ChantaO = 1ull << 28,
+    f_ChantaC = 1ull << 29,
+    f_SanshokuSeqO = 1ull << 30,
+    f_SanshokuSeqC = 1ull << 31,
+    f_SanshokuSet = 1ull << 32,
+    f_IttsuO = 1ull << 33,
+    f_IttsuC = 1ull << 34,
+    f_Toitoi = 1ull << 35,
+    f_Sanankou = 1ull << 36,
+    f_Sankantsu = 1ull << 37,
+    f_Chitoitsu = 1ull << 38,
+    f_Honroutou = 1ull << 39,
+    f_Shousangen = 1ull << 40,
+    f_HonitsuO = 1ull << 41,
+    f_HonitsuC = 1ull << 42,
+    f_JunchanO = 1ull << 43,
+    f_JunchanC = 1ull << 44,
+    f_Ryanpeikou = 1ull << 45,
+    f_ChinitsuO = 1ull << 46,
+    f_ChinitsuC = 1ull << 47,
+
+    f_ChuurenPoutou = 1ull << 48,
+    f_SuuAnkou = 1ull << 49,
+    f_SuuKantsu = 1ull << 50,
+    f_Daisangen = 1ull << 51,
+    f_Shosushi = 1ull << 52,
+    f_Daisushi = 1ull << 53,
+    f_Kokushi = 1ull << 54,
+    f_Ryuisou = 1ull << 55,
+    f_Chinroutou = 1ull << 56,
+    f_Tenhou = 1ull << 57,
+    f_Chihou = 1ull << 58,
+    f_DoubleYakuman = 1ull << 59,
+    f_Renhou = 1ull << 60
+};
+
+
 constexpr U64 f_FuMask = 0xff;
 constexpr U64 f_YakuMask = ~f_FuMask;
 
-constexpr U64 f_Riichi1         = 0x0000000000000100; // single or double
-constexpr U64 f_Riichi2         = 0x0000000000000200; // double only
 constexpr U64 f_Riichi          = f_Riichi1 | f_Riichi2;
-constexpr U64 f_Ippatsu         = 0x0000000000000400;
-constexpr U64 f_MenTsumo        = 0x0000000000000800;
-constexpr U64 f_Pinfu           = 0x0000000000001000;
-constexpr U64 f_Ipeikou         = 0x0000000000002000;
-constexpr U64 f_Haitei          = 0x0000000000004000;
-constexpr U64 f_Houtei          = 0x0000000000008000;
-constexpr U64 f_Rinshan         = 0x0000000000010000;
-constexpr U64 f_Chankan         = 0x0000000000020000;
-constexpr U64 f_Tanyao          = 0x0000000000040000;
-
-constexpr U64 f_Hatsu           = 0x0000000000080000;
-constexpr U64 f_Chun            = 0x0000000000100000;
-constexpr U64 f_Haku            = 0x0000000000200000;
-constexpr U64 f_EastP           = 0x0000000000400000;
-constexpr U64 f_EastS           = 0x0000000000800000;
 constexpr U64 f_East            = f_EastP | f_EastS;
-constexpr U64 f_SouthP          = 0x0000000001000000;
-constexpr U64 f_SouthS          = 0x0000000002000000;
 constexpr U64 f_South           = f_SouthP | f_SouthS;
-constexpr U64 f_WestS           = 0x0000000004000000;
 constexpr U64 f_West            = f_WestS;
-constexpr U64 f_NorthS          = 0x0000000008000000;
 constexpr U64 f_North           = f_NorthS;
-
-constexpr U64 f_ChantaO         = 0x0000000010000000;
-constexpr U64 f_ChantaC         = 0x0000000020000000;
 constexpr U64 f_Chanta          = f_ChantaO | f_ChantaC;
-constexpr U64 f_SanshokuSeqO    = 0x0000000040000000;
-constexpr U64 f_SanshokuSeqC    = 0x0000000080000000;
 constexpr U64 f_SanshokuSeq     = f_SanshokuSeqO | f_SanshokuSeqC;
-constexpr U64 f_SanshokuSet     = 0x0000000100000000;
-constexpr U64 f_IttsuO          = 0x0000000200000000;
-constexpr U64 f_IttsuC          = 0x0000000400000000;
-constexpr U64 f_Ittsu           = 0x0000000600000000;
-constexpr U64 f_Toitoi          = 0x0000000800000000;
-constexpr U64 f_Sanankou        = 0x0000001000000000;
-constexpr U64 f_Sankantsu       = 0x0000002000000000;
-constexpr U64 f_Chitoitsu       = 0x0000004000000000;
-constexpr U64 f_Honroutou       = 0x0000008000000000;
-constexpr U64 f_Shousangen      = 0x0000010000000000;
-constexpr U64 f_HonitsuO        = 0x0000020000000000;
-constexpr U64 f_HonitsuC        = 0x0000040000000000;
+constexpr U64 f_Ittsu           = f_IttsuO | f_IttsuC;
 constexpr U64 f_Honitsu         = f_HonitsuO | f_HonitsuC;
-constexpr U64 f_JunchanO        = 0x0000080000000000;
-constexpr U64 f_JunchanC        = 0x0000100000000000;
 constexpr U64 f_Junchan         = f_JunchanO | f_JunchanC;
-constexpr U64 f_Ryanpeikou      = 0x0000200000000000;
-constexpr U64 f_ChinitsuO       = 0x0000400000000000;
-constexpr U64 f_ChinitsuC       = 0x0000800000000000;
 constexpr U64 f_Chinitsu        = f_ChinitsuO | f_ChinitsuC;
-
-constexpr U64 f_ChuurenPoutou   = 0x0004000000000000;
-constexpr U64 f_SuuAnkou        = 0x0008000000000000;
-constexpr U64 f_SuuKantsu       = 0x0010000000000000;
-constexpr U64 f_Daisangen       = 0x0020000000000000;
-constexpr U64 f_Shosushi        = 0x0040000000000000;
-constexpr U64 f_Daisushi        = 0x0080000000000000;
-constexpr U64 f_Kokushi         = 0x0100000000000000;
-constexpr U64 f_Ryuisou         = 0x0200000000000000;
-constexpr U64 f_Chinroutou      = 0x0400000000000000;
-constexpr U64 f_Tenhou          = 0x0800000000000000;
-constexpr U64 f_Chihou          = 0x1000000000000000;
-constexpr U64 f_DoubleYakuman   = 0x2000000000000000;
-constexpr U64 f_YakumanMask     = 4610560118520545280ULL;
-
-constexpr U64 f_Renhou          = 0x4000000000000000;
 
 constexpr U64 f_OpenHandMask    = ~(f_Riichi | 
                                     f_Ippatsu | 
@@ -148,6 +149,18 @@ constexpr U64 f_Yaku3Fan        =   f_HonitsuC |
 constexpr U64 f_Yaku5Fan        =   f_ChinitsuO;
 constexpr U64 f_Yaku6Fan        =   f_ChinitsuC;
 
+constexpr U64 f_YakumanMask     =   f_ChuurenPoutou |
+                                    f_SuuAnkou |
+                                    f_SuuKantsu |
+                                    f_Daisangen |
+                                    f_Shosushi |
+                                    f_Daisushi |
+                                    f_Kokushi |
+                                    f_Ryuisou |
+                                    f_Chinroutou |
+                                    f_Tenhou |
+                                    f_Chihou;
+
 constexpr std::array<U64, 6> k_YakuVal 
 { f_Yaku1Fan, f_Yaku2Fan, f_Yaku3Fan, 0, f_Yaku5Fan, f_Yaku6Fan };
 
@@ -192,7 +205,7 @@ inline constexpr U16f basic_score(U64 yakus, U8f doras=0)
 }
 
 /**
- * Evaluate a particular yaku 
+ * Evaluate a particular yaku. Implemented for all single 
  * 
  * @tparam yaku The yaku to evaluate
  * @param combo The scoring combo reference to aggregate the result
@@ -203,82 +216,6 @@ inline constexpr U16f basic_score(U64 yakus, U8f doras=0)
  */
 template<U64 yaku>
 inline constexpr U64 eval(const Hand &hand, const Win &win, Tile agari_pai);
-
-template<>
-inline constexpr U64 eval<f_Pinfu>(const Hand &hand, const Win &win, Tile agari_pai)
-{
-    auto win_ptr = std::get_if<NormalWin>(&win);
-    if (win_ptr == nullptr)
-        return 25;
-    auto &[melds, pair] = *win_ptr;
-    U8f agari34 = agari_pai.id34();
-    U8f fu = k_BaseFu;
-    U8f wait_fu = 2;
-    for (const auto &meld : melds) // closed or semi-closed melds
-    {
-        if (_is_set(meld))
-        {
-            U8f triple_pts = 2;
-            if (_is_19(_int(meld)))
-                triple_pts <<= 1;
-
-            // meld must be closed unless the third tile is won from ron
-            if (agari34==_int(meld) && agari_pai.player()!=hand[0].player() && hand.hand_4hot_melds()[_int(meld)]!=meld)
-            ;
-            else
-                triple_pts <<= 1;
-            fu += triple_pts;            
-        }
-        else if ((agari34==_int(meld) && agari_pai.num() != c_num(7)) || agari34==_int(meld)+2 && agari_pai.num() != c_num(3))
-            wait_fu = 0;
-    }
-    for (const auto &meld : hand.melds()) // open melds
-    {
-        if (meld.is_set())
-        {
-            U8f triple_pts = 2;
-            if (meld.fourth())
-                triple_pts <<= 2;
-            if (meld.first().is_19())
-                triple_pts <<= 1;
-            fu += triple_pts;
-        }
-    }
-
-    fu += wait_fu;
-    if (fu == k_BaseFu)
-    {
-        if (hand.check(Hand::f_Open))
-            return fu + 10;
-        else if (agari_pai.player() != hand[0].player())
-            fu += 10;
-        return fu | f_Pinfu;
-    }
-    if (agari_pai.player() == hand[0].player())
-        fu += 2;
-    if (!hand.check(Hand::f_Open))
-        fu += 10;
-
-    return 10 * ((fu + 9) / 10);
-}
-
-template<>
-inline constexpr U64 eval<f_Ipeikou>(const Hand &hand, const Win &win, Tile agari_pai)
-{
-    auto win_ptr = std::get_if<NormalWin>(&win);
-    if (win_ptr == nullptr)
-        return 0;
-    auto &[melds, pair] = *win_ptr;
-    for (const auto &meld : melds)
-    {
-        if (_is_run(meld))
-        {
-            
-        }
-    }
-}
-
-
 
 U32f score_hand(const Hand &hand, const Win &win, Tile agari_pai);
